Fill table scimessages in constructor
As discussed in https://github.com/larshp/abapOpenChecks/pull/692 and https://github.com/larshp/abapOpenChecks/issues/698, table scimessages should be filled in the constructor. This table is used to evaluate the texts for findings.

This commit includes the necessary changes in class zcl_aoc_super and adjusts a few checks.