Handle obsolete comparison operators
If the (obsolete in later releases) comparison operators ('<>','=>' and '=<') are not recognized the parsing of expressions does not work as expected.

Also, assert that children which are added are actually bound. The problem with the obsolete operators was discovered because check number 59 terminated when it tried to process an initial reference. The assert stops the program when the root cause error occurs,